Renaming StreamLoop in favor of ResilientStream. Renamed package typeclasses in favor of util.
