+++ core/src/main/scala/com/github/gvolpe/fs2rabbit/program/ConsumingProgram.scala
@@ -19,7 +19,7 @@ package com.github.gvolpe.fs2rabbit.program
 import cats.effect.Async
 import com.github.gvolpe.fs2rabbit.algebra.{AckerConsumer, Consuming}
 import com.github.gvolpe.fs2rabbit.model._
-import com.github.gvolpe.fs2rabbit.typeclasses.StreamEval
+import com.github.gvolpe.fs2rabbit.util.StreamEval
 import com.rabbitmq.client.Channel
 import fs2.Stream
