+++ json-circe/src/main/scala/com/github/gvolpe/fs2rabbit/json/Fs2JsonDecoder.scala
@@ -17,8 +17,8 @@
 package com.github.gvolpe.fs2rabbit.json
 
 import cats.effect.Sync
+import com.github.gvolpe.fs2rabbit.util.{Log, StreamEval}
 import com.github.gvolpe.fs2rabbit.model.{AmqpEnvelope, DeliveryTag}
-import com.github.gvolpe.fs2rabbit.typeclasses.{Log, StreamEval}
 import fs2.{Pipe, Stream}
 import io.circe.parser.decode
 import io.circe.{Decoder, Error}
