+++ site/src/main/tut/examples/sample-acker.md
@@ -16,7 +16,7 @@ import com.github.gvolpe.fs2rabbit.json.Fs2JsonEncoder
 import com.github.gvolpe.fs2rabbit.model.AckResult._
 import com.github.gvolpe.fs2rabbit.model.AmqpHeaderVal._
 import com.github.gvolpe.fs2rabbit.model._
-import com.github.gvolpe.fs2rabbit.typeclasses.StreamEval
+import com.github.gvolpe.fs2rabbit.util.StreamEval
 import fs2.{Pipe, Stream}
 
 import scala.concurrent.ExecutionContext
@@ -84,9 +84,9 @@ At the edge of out program we define our effect, `cats.effect.IO` in this case, 
 
 ```tut:book:silent
 import cats.effect.IO
-import com.github.gvolpe.fs2rabbit.StreamLoop
 import com.github.gvolpe.fs2rabbit.config.Fs2RabbitConfig
 import com.github.gvolpe.fs2rabbit.interpreter.Fs2Rabbit
+import com.github.gvolpe.fs2rabbit.resiliency.ResilientStream
 
 import scala.concurrent.ExecutionContext
 
@@ -105,7 +105,7 @@ object IOAckerConsumer extends IOApp {
 
   override def start(args: List[String]): IO[Unit] =
     Fs2Rabbit[IO](config).flatMap { implicit interpreter =>
-      StreamLoop.run(new AckerConsumerDemo[IO]().program)
+      ResilientStream.run(new AckerConsumerDemo[IO]().program)
     }
 }
 ```
